feat: refactor

+++ project/apps/blog/src/app/posts.service.ts
@@ -15,37 +15,103 @@ import {
   UpdatePostDto,
 } from './dto';
 import { BlogExceptions } from './constants';
+import {
+  CommonPostRdo,
+  CreateLinkPostRdo,
+  CreatePhotoPostRdo,
+  CreateQuotePostRdo,
+  CreateTextPostRdo,
+  CreateVideoPostRdo,
+} from './rdo';
+import { PostStatus, PostType } from '@project/core';
 
 @Injectable()
 export class PostsService {
   constructor(private readonly postsRepository: PostsRepository) {}
 
-  async createVideoPost(dto: CreateVideoPostDto, userId: string) {
-    return this.postsRepository.createPost(dto, userId);
+  async createVideoPost(
+    dto: CreateVideoPostDto,
+    userId: string
+  ): Promise<CreateVideoPostRdo> {
+    const response = await this.postsRepository.createPost(dto, userId);
+
+    return {
+      ...response,
+      type: PostType.VIDEO,
+      status: PostStatus.DRAFT,
+      title: response.title as string,
+      videoUrl: response.videoUrl as string,
+    };
   }
 
-  async createTextPost(dto: CreateTextPostDto, userId: string) {
-    return this.postsRepository.createPost(dto, userId);
+  async createTextPost(
+    dto: CreateTextPostDto,
+    userId: string
+  ): Promise<CreateTextPostRdo> {
+    const response = await this.postsRepository.createPost(dto, userId);
+
+    return {
+      ...response,
+      type: PostType.TEXT,
+      status: PostStatus.DRAFT,
+      preview: response.preview as string,
+      title: response.title as string,
+      content: response.content as string,
+    };
   }
 
-  async createQuotePost(dto: CreateQuotePostDto, userId: string) {
-    return this.postsRepository.createPost(dto, userId);
+  async createQuotePost(
+    dto: CreateQuotePostDto,
+    userId: string
+  ): Promise<CreateQuotePostRdo> {
+    const response = await this.postsRepository.createPost(dto, userId);
+
+    return {
+      ...response,
+      type: PostType.QUOTE,
+      status: PostStatus.DRAFT,
+      quoteText: response.quoteText as string,
+      quoteAuthor: response.quoteAuthor as string,
+    };
   }
 
-  async createPhotoPost(dto: CreatePhotoPostDto, userId: string) {
-    return this.postsRepository.createPost(dto, userId);
+  async createPhotoPost(
+    dto: CreatePhotoPostDto,
+    userId: string
+  ): Promise<CreatePhotoPostRdo> {
+    const response = await this.postsRepository.createPost(dto, userId);
+
+    return {
+      ...response,
+      type: PostType.PHOTO,
+      status: PostStatus.DRAFT,
+      photoUrl: response.photoUrl as string,
+    };
   }
 
-  async createLinkPost(dto: CreateLinkPostDto, userId: string) {
-    return this.postsRepository.createPost(dto, userId);
+  async createLinkPost(
+    dto: CreateLinkPostDto,
+    userId: string
+  ): Promise<CreateLinkPostRdo> {
+    const response = await this.postsRepository.createPost(dto, userId);
+
+    return {
+      ...response,
+      type: PostType.LINK,
+      status: PostStatus.DRAFT,
+      link: response.link as string,
+      description: response.description as string,
+    };
   }
 
   async getPosts(dto: GetPostsDto) {
     return this.postsRepository.findManyPosts(dto);
   }
 
-  async getMyDrafts(userId: string) {
-    return this.postsRepository.findDraftsPosts(userId);
+  async getMyDrafts(userId: string): Promise<CommonPostRdo[]> {
+    return (await this.postsRepository.findDraftsPosts(
+      userId
+    )) as unknown as CommonPostRdo[];
   }
 
   async getPostById(id: string) {
@@ -80,11 +146,31 @@ export class PostsService {
     );
   }
 
-  async updatePost(dto: UpdatePostDto, postId: string, userId?: string) {
-    return this.postsRepository.updatePost(dto, postId, userId);
+  async updatePost(
+    dto: UpdatePostDto,
+    postId: string,
+    userId?: string
+  ): Promise<CommonPostRdo> {
+    const post = await this.postsRepository.findPostById(postId);
+
+    if (!post) {
+      throw new NotFoundException(BlogExceptions.POST_NOT_FOUND);
+    }
+
+    return (await this.postsRepository.updatePost(
+      dto,
+      postId,
+      userId
+    )) as unknown as CommonPostRdo;
   }
 
   async deletePost(postId: string, userId: string) {
+    const post = await this.postsRepository.findPostById(postId);
+
+    if (!post) {
+      throw new NotFoundException(BlogExceptions.POST_NOT_FOUND);
+    }
+
     return this.postsRepository.deletePost(postId, userId);
   }
 
@@ -94,6 +180,12 @@ export class PostsService {
 
   async likePost(postId: string, userId: string) {
     try {
+      const post = await this.postsRepository.findPostById(postId);
+
+      if (!post) {
+        throw new NotFoundException(BlogExceptions.POST_NOT_FOUND);
+      }
+
       const existingLike = await this.postsRepository.findLikeByUserAndPostId(
         postId,
         userId
@@ -102,12 +194,6 @@ export class PostsService {
       const like = existingLike
         ? await this.postsRepository.deleteLike(existingLike.id, userId)
         : await this.postsRepository.createLike(postId, userId);
-
-      const post = await this.postsRepository.findPostById(postId);
-
-      if (!post) {
-        throw new NotFoundException('Post not found');
-      }
 
       await this.updatePost(
         {
@@ -120,7 +206,7 @@ export class PostsService {
 
       return like;
     } catch {
-      throw new BadRequestException('Like failed');
+      throw new BadRequestException(BlogExceptions.LIKE_ACTION_FAILED);
     }
   }
 }
